Move dashboard-specific sagas into submodules.
